move robot model viz to mbot_bringup
