+++ mbot_bringup/launch/mbot_viz.launch.py
@@ -1,0 +1,39 @@
+#!/usr/bin/env python3
+"""
+Combined launch file for visualization with RViz.
+This launches both the robot state publisher and RViz together.
+"""
+
+import os
+from ament_index_python.packages import get_package_share_directory
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument
+from launch.substitutions import LaunchConfiguration
+from launch_ros.actions import Node
+import xacro
+
+def generate_launch_description():
+    # Launch Arguments
+    use_sim_time = LaunchConfiguration('use_sim_time')
+    rviz_config = os.path.join(get_package_share_directory('mbot_bringup'), 'rviz', 'mbot_viz.rviz')
+    
+    # Launch Arguments
+    declare_use_sim_time_cmd = DeclareLaunchArgument(
+        'use_sim_time',
+        default_value='false',
+        description='Use simulation clock instead of system clock')
+    
+    # RViz
+    rviz_node = Node(
+        package='rviz2',
+        executable='rviz2',
+        name='rviz2',
+        arguments=['-d', rviz_config],
+        parameters=[{'use_sim_time': use_sim_time}],
+        output='screen'
+    )
+    
+    return LaunchDescription([
+        declare_use_sim_time_cmd,
+        rviz_node
+    ])
