+++ mbot_description/README.md
@@ -1,7 +1,2 @@
 # mbot_description
 - This is where we store urdf files.
-- To launch the robot model and tf frames, and visualize it in the Rviz
-    ```bash
-    ros2 launch mbot_description rviz_launch.py
-    ```
-    - If you have launched the lidar, you can also visualize lidar scan on Rviz
